[Improvement] Changed some topic names to make the purpose more clear
The ESP publish topic is renamed from "esp_update" to "esp_to_app" and the Android publish topic vice versa.

Also the READY message now includes the sender to clarify it a bit more.

The connection with the MQTT broker is now established when the QR code has been successfully scanned.